% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_seqspec.R
\name{b_sepFasta}
\alias{b_sepFasta}
\title{Seperate sequences (fasta) from \code{bold_seqspec} output.}
\usage{
b_sepFasta(x, format = "tsv")
}
\arguments{
\item{x}{(object) The output of a \code{bold_seqspec} call.}

\item{format}{(character) The format used in the \code{bold_seqspec} call. One of 'tsv' (default) or 'xml'.}
}
\value{
A list of length two : the specimen data and the sequences list.
}
\description{
Seperate sequences (fasta) from \code{bold_seqspec} output.
}
\examples{
\dontrun{
res <- bold_seqspec(taxon='Osmia')
res <- b_sepFasta(res)
# (same as bold_seqspec(taxon='Osmia', sepFasta = TRUE))
}
}
