% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{bold-package}
\alias{bold}
\alias{bold-package}
\title{bold: A programmatic interface to the Barcode of Life data.}
\description{
bold: A programmatic interface to the Barcode of Life data.
}
\section{About}{


This package gives you access to data from GBIF \url{http://www.boldsystems.org/} via their API.
}

\section{Functions}{


There are currently two main functions.

\itemize{
  bold_specimens - Search for specimen data.
  bold_seq - Search for and retrieve sequences.
}

Interestingly, they provide xml and tsv format data for the specimen data, while they provide
fasta data format for the sequence data. So for the specimen data you can get back raw XML, or
a data frame parsed from the tsv data, while for sequence data you get back a list (b/c
sequences are quite long and would make a data frame unwieldy).
}

