% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_identify.R
\name{bold_identify_async}
\alias{bold_identify_async}
\title{async identify}
\usage{
bold_identify_async(x, db = "COX1", ...)
}
\description{
async identify
}
\examples{
res <- bold_identify_async(x = sequences)
}
