% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_identify_parents.R
\name{bold_identify_parents-deprecated}
\alias{bold_identify_parents-deprecated}
\title{Add taxonomic parent names to a data.frame}
\arguments{
\item{x}{(data.frame/list) list of data.frames - the output from a call to
\code{\link{bold_identify}}. or a single data.frame from the output from same.
required.}

\item{wide}{(logical) output in long or wide format. See Details.
Default: \code{FALSE}}

\item{taxid}{(character) A taxid name. Optional. See \code{Filtering} below.}

\item{taxon}{(character) A taxon name. Optional. See \code{Filtering} below.}

\item{tax_rank}{(character) A tax_rank name. Optional. See \code{Filtering}
below.}

\item{tax_division}{(character) A tax_division name. Optional. See
\code{Filtering} below.}

\item{parentid}{(character) A parentid name. Optional. See \code{Filtering}
below.}

\item{parentname}{(character) A parentname name. Optional. See \code{Filtering}
below.}

\item{taxonrep}{(character) A taxonrep name. Optional. See \code{Filtering}
below.}

\item{specimenrecords}{(character) A specimenrecords name. Optional.
See \code{Filtering} below.}

\item{...}{Further args passed on to \code{\link[crul]{verb-GET}}, main
purpose being curl debugging}
}
\description{
Add taxonomic parent names to a data.frame
}
\seealso{
\code{\link{bold-deprecated}}
}
\keyword{internal}
