% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_identify_parents.R
\name{bold_identify_parents}
\alias{bold_identify_parents}
\title{Add taxonomic parent names to a data.frame}
\usage{
bold_identify_parents(x)
}
\arguments{
\item{x}{(data.frame/list) list of data.frames - the output from a call to 
\code{\link{bold_identify}}. or a single data.frame from the output from
same. required.}
}
\value{
a list of the same length as the input
}
\description{
Add taxonomic parent names to a data.frame
}
\details{
This function gets unique set of taxonomic names from the input
data.frame, then queries \code{\link{bold_tax_name}} to get the 
taxonomic ID, passing it to \code{\link{bold_tax_id}} to get the parent
names, then attaches those to the input data.
}
\examples{
\dontrun{
x <- df <- bold_identify(sequences = sequences)
out <- bold_identify_parents(df)
str(out)
head(out$seq1)
}
}

