% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bold_seq}
\alias{bold_seq}
\title{Search BOLD for sequences.}
\usage{
bold_seq(taxon = NULL, ids = NULL, bin = NULL, container = NULL,
  institutions = NULL, researchers = NULL, geo = NULL, marker = NULL,
  callopts = list())
}
\arguments{
  \item{taxon}{(character) Returns all records containing
  matching taxa. Taxa includes the ranks of phylum, class,
  order, family, subfamily, genus, and species.}

  \item{ids}{(character) Returns all records containing
  matching IDs. IDs include Sample IDs, Process IDs, Museum
  IDs and Field IDs.}

  \item{bin}{(character) Returns all records contained in
  matching BINs. A BIN is defined by a Barcode Index Number
  URI.}

  \item{container}{(character) Returns all records
  contained in matching projects or datasets. Containers
  include project codes and dataset codes}

  \item{institutions}{(character) Returns all records
  stored in matching institutions. Institutions are the
  Specimen Storing Site. (spaces need to be encoded)}

  \item{researchers}{(character) Returns all records
  containing matching researcher names. Researchers include
  collectors and specimen identifiers. (spaces need to be
  encoded)}

  \item{geo}{(character) Returns all records collected in
  matching geographic sites. Geographic sites includes
  countries and province/states. (spaces need to be
  encoded)}

  \item{marker}{(character) Returns all records containing
  matching marker codes.}

  \item{callopts}{(character) curl debugging opts passed on
  to httr::GET}
}
\value{
A list with each element of length 4 with slots for id, name, gene, and sequence.
}
\description{
Get sequences for a taxonomic group
}
\examples{
\dontrun{
bold_seq(taxon='Coelioxys')
bold_seq(taxon='Aglae')
bold_seq(id='ACRJP618-11')
bold_seq(bin='BOLD:AAA5125')
bold_seq(container='ACRJP')
bold_seq(institutions='Biodiversity Institute of Ontario')
bold_seq(researchers='Thibaud Decaens')
bold_seq(geo='Canada')
bold_seq(marker='COI')
}
}

