% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bold_specimens}
\alias{bold_specimens}
\title{Search BOLD for specimens.}
\usage{
bold_specimens(taxon = NULL, geo = NULL, format = "xml",
  callopts = list())
}
\arguments{
  \item{taxon}{scientitic name of taxon, as e.g., "Cottus"
  (character)}

  \item{geo}{geographic area to exclude (character)}

  \item{format}{One of xml or tsv (character)}
}
\description{
Search BOLD for specimens.
}
\examples{
\dontrun{
bold_specimens(taxon='Osmia', geo='Costa Rica')
bold_specimens(taxon="Formicidae", geo="Canada")

res <- bold_specimens(taxon="Formicidae", geo="Canada")
head(res)
}
}

