% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_tax_id.R
\name{bold_tax_id-deprecated}
\alias{bold_tax_id-deprecated}
\title{Search BOLD for taxonomy data by BOLD ID.}
\arguments{
\item{id}{(integer) One or more BOLD taxonomic identifiers. required.}

\item{dataTypes}{(character) Specifies the datatypes that will be
returned. 'all' returns all data. 'basic' returns basic taxon information.
'images' returns specimen images.}

\item{includeTree}{(logical) If \code{TRUE} (default: \code{FALSE}), returns a list
containing information for parent taxa as well as the specified taxon.}

\item{response}{(logical) Default : FALSE. If TRUE, returns the object from
the Curl call. Useful for debugging and getting more detailed info on the
API call.}

\item{...}{Further args passed on to
\code{\link[crul:verb-GET]{crul::verb-GET}}, main purpose being curl
debugging}
}
\description{
Search BOLD for taxonomy data by BOLD ID.
}
\seealso{
\code{\link{bold-deprecated}}
}
\keyword{internal}
