% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getsampleids}
\alias{getsampleids}
\title{getseqs}
\usage{
getsampleids(taxoninc = NA, taxonexc = NA, geoinc = NA, geoexc = NA,
  url = "http://services.boldsystems.org/eSearch.php?", ...)
}
\arguments{
  \item{taxoninc}{taxonomic group to include in search
  (character)}

  \item{taxonexc}{taxonomic group to exclude in search
  (character)}

  \item{geoinc}{geographic location to include (character)}

  \item{geoexc}{geographic location to exclude (character)}

  \item{url}{base URL for the BOLD API (leave to default)}
}
\description{
Get sequences for a taxonomic group
}
\examples{
\dontrun{
getsampleids(taxoninc = 'Andrena')
getsampleids(taxoninc = 'Andrenidae')
}
}

