% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold-package.R
\docType{data}
\name{rank_ref}
\alias{rank_ref}
\title{Vector of bold taxonomic ranks}
\format{
named character vector of 10 rank names, named with their plural form
}
\description{
Vector of bold taxonomic ranks
}
\details{
This character vector is used to do data sorting/filtering based on the ordering
of ranks and to convert a rank's plural form to singular.
}
\keyword{data}
