% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{sequences}
\alias{sequences}
\title{List of 3 nucleotide sequences to use in examples for the bold_identify function}
\description{
List of 3 nucleotide sequences to use in examples for the bold_identify function
}
\keyword{data}

