% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{bold-package}
\alias{bold}
\alias{bold-package}
\title{bold: A programmatic interface to the Barcode of Life data.}
\description{
bold: A programmatic interface to the Barcode of Life data.
}
\section{About}{
  This package gives you access to data from GBIF
  \url{http://www.boldsystems.org/} via their API.
}

\section{Functions}{
  There are currently two main functions.

  \itemize{ bold_specimens - Search for specimen data.
  bold_seq - Search for and retrieve sequences. }
}

