% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bold_identify_taxonomy.R
\name{bold_identify_taxonomy}
\alias{bold_identify_taxonomy}
\title{Add taxonomic parent names to a data.frame}
\usage{
bold_identify_taxonomy(x, taxOnly = TRUE)
}
\arguments{
\item{x}{(data.frame|list) A single data.frame or a list of - the output from
a call to \code{\link{bold_identify}}. Required.}

\item{taxOnly}{(logical) If TRUE, only the taxonomic names and ids are added (equivalent format to the results of \code{\link{bold_identify_parents}} when \code{wide}is set to TRUE. If FALSE, also joins the rest of the data returned by \code{\link{bold_specimens}}.}

\item{response}{(logical) Default : FALSE. If TRUE, returns the object from
the Curl call. Useful for debugging and getting more detailed info on the
API call.}

\item{...}{Further args passed on to
\code{\link[crul:verb-GET]{crul::verb-GET}}, main purpose being curl
debugging}
}
\value{
a data.frame or a list of data.frames with added taxonomic
classification.
}
\description{
Add taxonomic parent names to a data.frame
}
\details{
This function gets the process ids from the
input data.frame(s) (ID column), then queries \code{\link{bold_specimens}}
to get the sample information and cbinds it to the input data.frame(s).

Records in the input data that do not have matches for parent names
simply get NA values in the added columns.
}
\examples{
\dontrun{
seqs <- bold_identify(sequences = bold::sequences$seq2)

seqs_tax <- bold_identify_taxonomy(seqs)
head(seqs_tax[[1]])

x <- bold_seq(taxon = "Satyrium")
seqs <- bold_identify(x$sequence[1:2])
seqs_tax <- bold_identify_taxonomy(seqs)
seqs_tax
}

}
