% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bold-package.R
\docType{data}
\name{sequences}
\alias{sequences}
\title{List of 3 nucleotide sequences to use in examples for the
\code{\link{bold_identify}} function}
\description{
List of 3 nucleotide sequences to use in examples for the
\code{\link{bold_identify}} function
}
\details{
Each sequence is a character string, of lengths 410, 600, and 696.
}
\keyword{data}

