% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bold-package.R
\docType{data}
\name{sequences}
\alias{sequences}
\title{List of 3 nucleotide sequences to use in examples for the bold_identify function}
\description{
List of 3 nucleotide sequences to use in examples for the bold_identify function
}
\keyword{data}

